PROJECT/black/requirements.txt
# ============================================================================
# requirements.txt · EntroDiff
# ----------------------------------------------------------------------------
# 锁定的运行时依赖。版本上限留出 minor patch 弹性。
# 开发依赖见 pyproject.toml [project.optional-dependencies]
# ============================================================================

# --- 数值核心 ---
numpy>=1.24,<2.0
scipy>=1.11,<2.0

# --- 深度学习 ---
torch>=2.2,<3.0
einops>=0.7,<1.0

# --- 实验配置 / IO ---
PyYAML>=6.0,<7.0
hydra-core>=1.3,<2.0
omegaconf>=2.3,<3.0

# --- 可视化 ---
matplotlib>=3.8,<4.0
tqdm>=4.66,<5.0

# --- 最优传输（用于 W_1 / W_2 估计）---
POT>=0.9,<1.0
